pyproject.toml
[project]
name = "mcp-google-workspace"
version = "0.1.0"
description = "MCP server for Google Workspace (Gmail, Drive, Sheets) with auto OAuth flow and Gmail signature support"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Sensio.cz", email = "info@sensio.cz"}]
readme = "README.md"
keywords = ["mcp", "google", "gmail", "drive", "sheets", "workspace"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/Sensio-cz/mcp-google-workspace"
Issues = "https://github.com/Sensio-cz/mcp-google-workspace/issues"

[project.scripts]
mcp-google-workspace = "mcp_google_workspace.__main__:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mcp_google_workspace*"]
